--- src/main/kotlin/com/arabyte/arabyteapi/SecurityConfiguration.kt
@@ -1,24 +1,0 @@
-package com.arabyte.arabyteapi
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-import org.springframework.security.web.SecurityFilterChain
-
-@EnableWebSecurity
-@Configuration
-class SecurityConfiguration {
-    @Bean
-    fun securityFilterChain(http: HttpSecurity): SecurityFilterChain {
-        http
-            .csrf { it.disable() }
-            .authorizeHttpRequests {
-                it.anyRequest().permitAll()
-            }
-            .formLogin { it.disable() }
-            .httpBasic { it.disable() }
-
-        return http.build()
-    }
-}
