--- src/main/kotlin/com/arabyte/arabyteapi/auth/dto/KakaoTokenResponse.kt
@@ -1,11 +1,0 @@
-package com.arabyte.arabyteapi.auth.dto
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class KakaoTokenResponse(
-    @JsonProperty("access_token") val accessToken: String,
-    @JsonProperty("token_type") val tokenType: String,
-    @JsonProperty("refresh_token") val refreshToken: String,
-    @JsonProperty("expires_in") val expiresIn: Int,
-    @JsonProperty("scope") val scope: String
-)
