--- src/main/kotlin/com/arabyte/arabyteapi/auth/dto/RefreshAccessTokenRequestBody.kt
@@ -1,5 +1,0 @@
-package com.arabyte.arabyteapi.auth.dto
-
-class RefreshAccessTokenRequestBody(
-    val refreshToken: String
-)
