--- src/main/kotlin/com/arabyte/arabyteapi/auth/dto/RefreshAccessTokenResponse.kt
@@ -1,5 +1,0 @@
-package com.arabyte.arabyteapi.auth.dto
-
-class RefreshAccessTokenResponse(
-    val accessToken: String
-)
