--- src/main/kotlin/com/arabyte/arabyteapi/auth/dto/TokenWithUserResponse.kt
@@ -1,9 +1,0 @@
-package com.arabyte.arabyteapi.auth.dto
-
-import com.arabyte.arabyteapi.user.entitiy.User
-
-class TokenWithUserResponse(
-    val accessToken: String,
-    val refreshToken: String,
-    val user: User
-)
