--- src/main/kotlin/com/arabyte/arabyteapi/common/entity/BaseEntity.kt
@@ -1,21 +1,0 @@
-package com.arabyte.arabyteapi.common.entity
-
-import jakarta.persistence.*
-import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
-import org.springframework.data.jpa.domain.support.AuditingEntityListener
-import java.time.LocalDateTime
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener::class)
-class BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long = 0
-
-    @CreatedDate
-    var createdAt: LocalDateTime? = LocalDateTime.MIN
-
-    @LastModifiedDate
-    var updatedAt: LocalDateTime? = LocalDateTime.MIN
-}
