--- src/main/kotlin/com/arabyte/arabyteapi/common/enums/JobCategory.kt
@@ -1,5 +1,0 @@
-package com.arabyte.arabyteapi.common.enums
-
-enum class JobCategory {
-    RESTAURANT, RETAIL, SERVICE, SALES, MANUFACTURING, IT, DESIGN, OFFICE, EDUCATION, DELIVERY, MEDIA, HOSPITAL
-}
