--- src/main/kotlin/com/arabyte/arabyteapi/common/exception/CustomException.kt
@@ -1,6 +1,0 @@
-package com.arabyte.arabyteapi.common.exception
-
-import org.springframework.web.server.ResponseStatusException
-
-class CustomException(customError: CustomError) :
-    ResponseStatusException(customError.status, customError.message)
