+++ src/main/kotlin/com/arabyte/arabyteapi/domain/article/ArticleImage.kt
@@ -1,0 +1,10 @@
+package com.arabyte.arabyteapi.domain.article
+
+import com.arabyte.arabyteapi.global.entity.BaseEntity
+import jakarta.persistence.Entity
+
+@Entity
+class ArticleImage(
+    val articleId: Long,
+    val url: String
+) : BaseEntity()
