+++ src/main/kotlin/com/arabyte/arabyteapi/domain/article/ArticleKind.kt
@@ -1,0 +1,11 @@
+package com.arabyte.arabyteapi.domain.article
+
+import com.arabyte.arabyteapi.global.entity.BaseEntity
+import jakarta.persistence.Entity
+
+@Entity
+class ArticleKind(
+    val name: String,
+    val description: String,
+    val useAnonymous: Boolean
+) : BaseEntity()
