+++ src/main/kotlin/com/arabyte/arabyteapi/domain/auth/controller/KakaoAuthController.kt
@@ -1,0 +1,63 @@
+package com.arabyte.arabyteapi.domain.auth.controller
+
+import com.arabyte.arabyteapi.domain.auth.dto.KakaoUserResponse
+import com.arabyte.arabyteapi.domain.auth.dto.RefreshAccessTokenRequestBody
+import com.arabyte.arabyteapi.domain.auth.dto.RefreshAccessTokenResponse
+import com.arabyte.arabyteapi.domain.auth.dto.TokenWithUserResponse
+import com.arabyte.arabyteapi.domain.auth.service.KakaoAuthService
+import com.arabyte.arabyteapi.domain.auth.util.JwtProvider
+import com.arabyte.arabyteapi.global.exception.InvalidTokenException
+import org.springframework.web.bind.annotation.*
+
+@RestController
+@RequestMapping("/auth/kakao")
+class KakaoAuthController(
+    private val kakaoAuthService: KakaoAuthService,
+    private val jwtProvider: JwtProvider,
+) {
+    @GetMapping("/callback")
+    fun kakaoLogin(
+        @RequestParam("code") code: String
+    ): TokenWithUserResponse {
+        val kakaoAccessToken = kakaoAuthService.getKakaoAccessToken(code)
+        val kakaoUser = kakaoAuthService.getKakaoUserInfo(kakaoAccessToken)
+
+        val user = kakaoAuthService.loginOrRegister(kakaoUser)
+        val accessToken = jwtProvider.generateAccessToken(user.id.toString())
+        val refreshToken = jwtProvider.generateRefreshToken(user.id.toString())
+
+        return TokenWithUserResponse(
+            accessToken,
+            refreshToken,
+            user
+        )
+    }
+
+    @PostMapping("/login")
+    fun loginOrRegister(
+        @RequestBody kakaoUser: KakaoUserResponse
+    ): TokenWithUserResponse {
+        val user = kakaoAuthService.loginOrRegister(kakaoUser)
+        val accessToken = jwtProvider.generateAccessToken(user.id.toString())
+        val refreshToken = jwtProvider.generateRefreshToken(user.id.toString())
+
+        return TokenWithUserResponse(
+            accessToken,
+            refreshToken,
+            user
+        )
+    }
+
+    @PostMapping("/refresh")
+    fun refreshAccessToken(
+        @RequestBody body: RefreshAccessTokenRequestBody
+    ): RefreshAccessTokenResponse {
+        if (!jwtProvider.isValidToken(body.refreshToken)) {
+            throw InvalidTokenException()
+        }
+
+        val userId = jwtProvider.getUserId(body.refreshToken)
+        val accessToken = jwtProvider.generateAccessToken(userId)
+        return RefreshAccessTokenResponse(accessToken)
+    }
+}
