+++ src/main/kotlin/com/arabyte/arabyteapi/domain/auth/dto/RefreshAccessTokenRequestBody.kt
@@ -1,0 +1,5 @@
+package com.arabyte.arabyteapi.domain.auth.dto
+
+class RefreshAccessTokenRequestBody(
+    val refreshToken: String
+)
