+++ src/main/kotlin/com/arabyte/arabyteapi/domain/auth/dto/RefreshAccessTokenResponse.kt
@@ -1,0 +1,5 @@
+package com.arabyte.arabyteapi.domain.auth.dto
+
+class RefreshAccessTokenResponse(
+    val accessToken: String
+)
