+++ src/main/kotlin/com/arabyte/arabyteapi/domain/auth/util/JwtAuthenticationFilter.kt
@@ -1,0 +1,36 @@
+package com.arabyte.arabyteapi.domain.auth.util
+
+import jakarta.servlet.FilterChain
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
+import org.springframework.security.core.context.SecurityContextHolder
+import org.springframework.security.core.userdetails.User
+import org.springframework.security.core.userdetails.UserDetails
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource
+import org.springframework.stereotype.Component
+import org.springframework.web.filter.OncePerRequestFilter
+
+@Component
+class JwtAuthenticationFilter(
+    private val jwtProvider: JwtProvider
+) : OncePerRequestFilter() {
+    override fun doFilterInternal(
+        request: HttpServletRequest,
+        response: HttpServletResponse,
+        filterChain: FilterChain
+    ) {
+        val token = jwtProvider.resolveToken(request)
+
+        if (!token.isNullOrBlank() && jwtProvider.isValidToken(token)) {
+            val userId = jwtProvider.getUserId(token)
+            val userDetails: UserDetails =
+                User.withUsername(userId).password("").authorities(emptyList()).build()
+            val authentication = JwtAuthenticationToken(userDetails, token, userDetails.authorities)
+            authentication.details = WebAuthenticationDetailsSource().buildDetails(request)
+
+            SecurityContextHolder.getContext().authentication = authentication
+        }
+
+        filterChain.doFilter(request, response)
+    }
+}
