+++ src/main/kotlin/com/arabyte/arabyteapi/domain/user/enums/JobCategory.kt
@@ -1,0 +1,5 @@
+package com.arabyte.arabyteapi.domain.user.enums
+
+enum class JobCategory {
+    RESTAURANT, RETAIL, SERVICE, SALES, MANUFACTURING, IT, DESIGN, OFFICE, EDUCATION, DELIVERY, MEDIA, HOSPITAL
+}
