+++ src/main/kotlin/com/arabyte/arabyteapi/domain/user/service/UserService.kt
@@ -1,0 +1,18 @@
+package com.arabyte.arabyteapi.domain.user.service
+
+import com.arabyte.arabyteapi.domain.user.entitiy.User
+import com.arabyte.arabyteapi.domain.user.repository.UserRepository
+import org.springframework.stereotype.Service
+
+@Service
+class UserService(
+    private val userRepository: UserRepository
+) {
+    fun getUserByKakaoId(id: Long): User? {
+        return userRepository.findByKakaoId(id)
+    }
+
+    fun saveUser(newUser: User): User {
+        return userRepository.save(newUser)
+    }
+}
