+++ src/main/kotlin/com/arabyte/arabyteapi/global/exception/CustomException.kt
@@ -1,0 +1,6 @@
+package com.arabyte.arabyteapi.global.exception
+
+import org.springframework.web.server.ResponseStatusException
+
+class CustomException(customError: CustomError) :
+    ResponseStatusException(customError.status, customError.message)
