--- src/main/kotlin/com/arabyte/arabyteapi/user/repository/UserRepository.kt
@@ -1,8 +1,0 @@
-package com.arabyte.arabyteapi.user.repository
-
-import com.arabyte.arabyteapi.user.entitiy.User
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface UserRepository : JpaRepository<User, Long> {
-    fun findByKakaoId(kakaoId: Long): User?
-}
